requirements.txt
# YOLOv8 Food Detection Model - Dependencies

ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
PyYAML>=6.0

# FastAPI Server Dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
